backend/requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-bootstrap==0.2.4
django-bootstrap3==24.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
google-api-core==2.19.0
google-api-python-client==2.131.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.0
gunicorn==22.0.0
httplib2==0.22.0
idna==3.7
oauth2client==4.1.3
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyDrive==1.3.1
pyparsing==3.1.2
PyYAML==6.0.1
requests==2.32.2
rsa==4.9
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
wheel==0.43.0
whitenoise==6.6.0
